+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+#### 1.2.0 - 2014-11-03
+ - Added new feature: stork-maven-stage (stages application in target/stage)
+ - Added new feature: stork-maven-assembly (creates tarball from target/stage)
+ - Added new feature: stork-play-assembly (creates launcher for play app and
+     also creates tarball for distribution)
+
 #### 1.1.0 - 2014-10-29
  - Renamed project to Stork.
  - Improved docs